Binding PFlags should not Overwrite Default Values
This patch updates the logic for the BindPFlags function. Because Viper
is the configuration hub for applications, it makes sense that it should
be the authority on configuration details, such as default values.
Therefore, in a scenario where a developer is initializing Viper with
default values for a given configuration, the default values should not
be overwritten later if PFlags are bound to the same configuration keys.
Instead, any explicit calls a developer makes to viper.SetDefault should
be the expected default value for a key.